users/requirements.txt
Authlib==1.3.0
blinker==1.8.1
cffi==1.16.0
click==8.1.7
cryptography==42.0.6
Flask==3.0.3
googleapis-common-protos==1.62.0
grpcio==1.62.2
grpcio-status==1.62.2
grpcio-tools==1.62.2
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
protobuf==4.25.3
psycopg2-binary==2.9.9
pycparser==2.22
PyYAML==6.0.1
setuptools==69.5.1
watchdog==4.0.0
Werkzeug==3.0.2
